requirements.txt
certifi==2018.8.13
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.13
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-decouple==3.1
pytz==2018.5
requests==2.19.1
uritemplate==3.0.0
urllib3==1.23
